Edit the schedule template of tvm to make it work when there are more than one reduction calculation in an output.
Edit fused ops files to set output datatypes identical to XLA.

+++ tests/operators/gpu/test_fused_bn_follow_relu.py
@@ -57,7 +57,7 @@ def gen_data(in_shape, in_dtype, inter_dtype, layout, out_dtype):
 
     return data, output, expect
 
-def test_fused_bn_follow_relu(in_shape, in_dtype='float16', layout='NHWC', out_dtype='float32', poly_sch=False):
+def test_fused_bn_follow_relu(in_shape, in_dtype='float16', layout='NHWC', out_dtype='float16', poly_sch=False):
 
     if layout != "NHWC" and layout != "NCHW":
         raise NotImplementedError(
